+++ pkg/server/app/doc.go
@@ -1,0 +1,4 @@
+/*
+Package app implements application business logic
+*/
+package app
