+++ pkg/server/database/models.go
@@ -131,6 +131,7 @@ type Digest struct {
 	UUID      string    `json:"uuid" gorm:"primary_key:true;type:uuid;index;default:uuid_generate_v4()"`
 	RuleID    int       `gorm:"index"`
 	UserID    int       `gorm:"index"`
+	Version   int       `gorm:"version"`
 	Notes     []Note    `gorm:"many2many:digest_notes;association_foreignKey:uuid;association_jointable_foreignkey:note_uuid;jointable_foreignkey:digest_uuid;"`
 	CreatedAt time.Time `json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`
