+++ pkg/server/operations/doc.go
@@ -1,0 +1,5 @@
+/*
+Package operations encapsulates operations which involves
+database and are not necessarily in the context of app.
+*/
+package operations
