+++ web/src/components/Header/AccountMenu.scss
@@ -108,6 +108,7 @@
 .email {
   font-weight: 600;
   white-space: normal;
+  word-break: break-all;
 }
 
 .session-notice-wrapper {
